refactor(LottoStarter) : MVC 모델 제거 및 이름 변경(LottoController)
