+++ src/main/java/lotto/Application.java
@@ -6,6 +6,6 @@ public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
         AppConfig appConfig = new AppConfig();
-        appConfig.getLottoController().run();
+        appConfig.getLottoStarter().run();
     }
 }
