+++ src/main/java/lotto/config/AppConfig.java
@@ -1,6 +1,6 @@
 package lotto.config;
 
-import lotto.controller.LottoController;
+import lotto.LottoStarter;
 import lotto.util.InputParser;
 import lotto.util.LottoNumberGenerator;
 import lotto.util.RandomLottoNumberGenerator;
@@ -12,7 +12,7 @@ public class AppConfig {
     private final OutputView outputView;
     private final LottoNumberGenerator lottoNumberGenerator;
     private final InputParser inputParser;
-    private final LottoController lottoController;
+    private final LottoStarter lottoStarter;
 
     private static final int LOTTO_NUMBER_COUNT = 6;
     private static final int LOTTO_NUMBER_MIN = 1;
@@ -22,11 +22,11 @@ public class AppConfig {
         this.outputView = new OutputView();
         this.lottoNumberGenerator = new RandomLottoNumberGenerator(LOTTO_NUMBER_COUNT, LOTTO_NUMBER_MIN, LOTTO_NUMBER_MAX);
         this.inputParser = new InputParser();
-        this.lottoController = new LottoController(inputView, outputView, lottoNumberGenerator, inputParser);
+        this.lottoStarter = new LottoStarter(inputView, outputView, lottoNumberGenerator, inputParser);
     }
 
-    public LottoController getLottoController() {
-        return lottoController;
+    public LottoStarter getLottoStarter() {
+        return lottoStarter;
     }
 
 }
